Adding Derpybot.py crash prevention
I had a crash when updating to 0.9.3.7 because I forgot to add the '\config\config.cfg' file. This piece of code should stop the program before it happens and tell the user why the program is stopped. 

The second try code does not stop the program because if '\modules\derpymarkov\config\config.cfg' is not present the program will not crash, instead it wont load any known text files it seems. It will simply inform the user something is not right.